refactor: improve logging format for 64-bit values across multiple components
- Updated logging statements in various files to use portable format specifiers (e.g., PRIx64, PRIu64) for 64-bit integers, enhancing consistency and readability.
- Adjusted error and info logs in Aarch64 and x86_64 architecture files, as well as in disk management and guest agent components, to ensure proper formatting of addresses and sizes.
- Included necessary header for <cinttypes> to support the new format specifiers.

+++ src/core/guest_agent/guest_agent_handler.cpp
@@ -103,7 +103,7 @@ void GuestAgentHandler::StartSyncHandshake() {
     oss << '\n';
 
     SendRaw(oss.str());
-    LOG_INFO("GuestAgent: sent guest-sync-delimited id=%lld", (long long)id);
+    LOG_INFO("GuestAgent: sent guest-sync-delimited id=%" PRId64, id);
 }
 
 void GuestAgentHandler::OnDataReceived(const uint8_t* data, size_t len) {
@@ -192,7 +192,7 @@ void GuestAgentHandler::SendCommand(const std::string& command) {
     oss << R"({"execute":")" << JsonEscape(command)
         << R"(","id":)" << id << "}\n";
 
-    LOG_INFO("GuestAgent: sending %s (id=%llu)", command.c_str(), (unsigned long long)id);
+    LOG_INFO("GuestAgent: sending %s (id=%" PRIu64 ")", command.c_str(), id);
     SendRaw(oss.str());
 }
 
@@ -209,7 +209,7 @@ void GuestAgentHandler::SendCommand(const std::string& command,
         << R"(","arguments":)" << arguments_json
         << R"(,"id":)" << id << "}\n";
 
-    LOG_INFO("GuestAgent: sending %s (id=%llu)", command.c_str(), (unsigned long long)id);
+    LOG_INFO("GuestAgent: sending %s (id=%" PRIu64 ")", command.c_str(), id);
     SendRaw(oss.str());
 }
